updated javadoc and moved Min's testing into packages
