+++ src/main/java/ca/unknown/bot/use_cases/schedule_reminder/ScheduledReminderInteractor.java
@@ -12,8 +12,8 @@ import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import java.util.*;
 
 /**
- * Primary Interactor for the scheduled reminder use case. It calls subordinate interactors to create events
- * and is tasked with setting up direct message reminders.
+ * Primary Interactor for the scheduled reminder use case. It calls subordinate interactors to create events, send
+ * delayed direct message reminders, and remove events once they have passed.
  */
 public class ScheduledReminderInteractor extends ListenerAdapter {
     /**
@@ -29,18 +29,26 @@ public class ScheduledReminderInteractor extends ListenerAdapter {
 
     /**
      * Performs the event interaction based on a slash command entered by the user.
+     * Upon scheduling an event:
+     * Loads the persisted repo into the cache upon restarting the program, and then checks to see
+     * if the user is scheduling an event for the first time, creating a new empty schedule for them if
+     * necessary. If the user tries to schedule an event in the past, it will prevent them from doing so.
+     * Subordinate interactors are called to schedule new events, send direct message reminders, and remove
+     * events once they have passed, adhering to the single responsibility principle.
+     * Upon trying to clear events:
+     * Checks if the user has an ongoing schedule to clear or if the specified event is actually in the
+     * schedule, preventing the user from proceeding if it is not. If the check passes, then the event(s) will be
+     * removed from the user's schedule, and they are unsubscribed from the corresponding reminder alert(s).
      * @param event represents a SlashCommandInteraction event.
      */
     @Override
     public void onSlashCommandInteraction(SlashCommandInteractionEvent event) {
         User discordUser = event.getUser();
 
-        // collects the user's discord username and userID
         String username = discordUser.getName();
 
         long userId = discordUser.getIdLong();
 
-         // loads the persisted repo onto the cache
         if(scheduleDAO.emptyCache("src/main/java/ca/unknown/bot/data_access/schedule_reminder/schedule_repository.json")){
             scheduleDAO.loadRepo("src/main/java/ca/unknown/bot/data_access/schedule_reminder/schedule_repository.json");
             new RestartScheduleInteractor(scheduleDAO, event.getJDA()).execute();
@@ -48,19 +56,15 @@ public class ScheduledReminderInteractor extends ListenerAdapter {
         }
 
         if (event.getName().equals("schedule-event")) {
-            // Tell discord we received the command, send a thinking... message to the user
             event.deferReply().queue();
 
-            // if the user is scheduling an event for the first time, make a new Schedule object for their use
             if(!scheduleDAO.existsByUser(username)){
                 scheduleDAO.saveNewUser(scheduleFactory.create(username, userId));
             }
 
-            // calls a controller to receive user input and convert it into data the subordinate interactor can use
             ScheduledReminderController scheduledReminderController = new ScheduledReminderController(event);
             ScheduledReminderInputData scheduledReminderInputData = scheduledReminderController.getScheduledReminderInputData();
 
-            // if the user tries to schedule an event in the past then stop execution
             if(scheduledReminderInputData.getEventDate().compareTo(new Date()) < 0){
                 event.getHook().sendMessage("You can only schedule a future event!").complete();
                 return;
@@ -71,29 +75,22 @@ public class ScheduledReminderInteractor extends ListenerAdapter {
                 return;
             }
 
-            // call subordinate interactor to schedule new generic event and return it
             ScheduledEvent newEvent = new ScheduledEventInteractor(scheduleDAO).execute(scheduledReminderInputData, username, event);
 
-            // call subordinate interactor to set up a delayed queue for sending a private message reminder to the user
             new SendReminderInteractor(scheduleDAO).execute(discordUser, newEvent);
 
-            // call subordinate interactor to clean up event from schedule after its passed
             new RemovePassedEventInteractor(scheduleDAO).execute(username, newEvent);
         }
         else if (event.getName().equals("schedule-exam")) {
-            // Tell discord we received the command, send a thinking... message to the user
             event.deferReply().queue();
 
-            // if the user is scheduling an event for the first time, make a new Schedule object for their use
             if(!scheduleDAO.existsByUser(username)){
                 scheduleDAO.saveNewUser(scheduleFactory.create(username, userId));
             }
 
-            // calls a controller to receive user input and convert it into data the subordinate interactor can use
             ScheduledReminderController scheduledReminderController = new ScheduledReminderController(event);
             ScheduledReminderInputData scheduledReminderInputData = scheduledReminderController.getScheduledReminderInputData();
 
-            // if the user tries to schedule an event in the past then stop execution
             if(scheduledReminderInputData.getEventDate().compareTo(new Date()) <= 0){
                 event.getHook().sendMessage("You can only schedule a future event!").complete();
                 return;
@@ -103,30 +100,23 @@ public class ScheduledReminderInteractor extends ListenerAdapter {
                 return;
             }
 
-            // call subordinate interactor to schedule new exam event and return it
             ScheduledEvent newExam = new ScheduledExamInteractor(scheduleDAO).execute(scheduledReminderInputData, username, event);
 
-            // call subordinate interactor to set up a delayed queue for sending a private message reminder to the user
             new SendReminderInteractor(scheduleDAO).execute(discordUser, newExam);
 
-            // call subordinate interactor to clean up event from schedule after its passed
             new RemovePassedEventInteractor(scheduleDAO).execute(username, newExam);
 
         }
         else if (event.getName().equals("schedule-assignment")){
-            // Tell discord we received the command, send a thinking... message to the user
             event.deferReply().queue();
 
-            // if the user is scheqduling an event for the first time, make a new Schedule object for their use
             if(!scheduleDAO.existsByUser(username)){
                 scheduleDAO.saveNewUser(scheduleFactory.create(username, userId));
             }
 
-            // calls a controller to receive user input and convert it into data the subordinate interactor can use
             ScheduledReminderController scheduledReminderController = new ScheduledReminderController(event);
             ScheduledReminderInputData scheduledReminderInputData = scheduledReminderController.getScheduledReminderInputData();
 
-            // if the user tries to schedule an event in the past then stop execution
             if(scheduledReminderInputData.getEventDate().compareTo(new Date()) < 0){
                 event.getHook().sendMessage("You can only schedule a future event!").complete();
                 return;
@@ -137,22 +127,15 @@ public class ScheduledReminderInteractor extends ListenerAdapter {
                 return;
             }
 
-            // call subordinate interactor to schedule new assignment event and return it
             ScheduledEvent newAssignment = new ScheduledAssignmentInteractor(scheduleDAO).execute(scheduledReminderInputData, username, event);
 
-            // call subordinate interactor to set up a delayed queue for sending a private message reminder to the user
             new SendReminderInteractor(scheduleDAO).execute(discordUser, newAssignment);
 
-
-            // call subordinate interactor to clean up event from schedule after its passed
             new RemovePassedEventInteractor(scheduleDAO).execute(username, newAssignment);
 
 
         }
         else if (event.getName().equals("current-schedule")){
-            // if the user has never scheduled an event before (so their schedule doesn't exist in the cache) or
-            // they have an existing schedule with no ongoing events, then alert the user that there are no events
-            // to be displayed
             if(!scheduleDAO.existsByUser(username)){
                 event.reply("You have no scheduled events.").queue();
             }
@@ -160,14 +143,11 @@ public class ScheduledReminderInteractor extends ListenerAdapter {
                 event.reply("You have no scheduled events.").queue();
             }
             else{
-                // otherwise displays a String representation of the user's ongoing event schedule
                 scheduleDAO.getSchedule(username).sort();
                 event.reply(scheduleDAO.getSchedule(username).toString()).queue();
             }
         }
         else if(event.getName().equals("clear-event")){
-            // if the user doesn't have an existing schedule in the cache, then alert them that there is no schedule
-            // to clear
             if(!scheduleDAO.existsByUser(username)){
                 event.reply("There is no ongoing schedule to clear.").queue();
             }
@@ -187,8 +167,6 @@ public class ScheduledReminderInteractor extends ListenerAdapter {
             }
         }
         else if(event.getName().equals("clear-schedule")){
-            // if the user doesn't have an existing schedule in the cache, then alert them that there is no schedule
-            // to clear
             if(!scheduleDAO.existsByUser(username)){
                 event.reply("There is no ongoing schedule to clear.").queue();
             }
@@ -196,16 +174,12 @@ public class ScheduledReminderInteractor extends ListenerAdapter {
                 event.reply("There is no ongoing schedule to clear.").queue();
             }
             else{
-                // otherwise, use the DAO to access the user's schedule and clear it
                 scheduleDAO.getSchedule(username).clearSched();
 
-                // unsubscribe user from all reminder alerts
                 scheduleDAO.removeAllChecks(username);
 
-                // update repo with the new cache containing the cleared schedule
                 scheduleDAO.saveToFile("src/main/java/ca/unknown/bot/data_access/schedule_reminder/schedule_repository.json");
 
-                // alert the user of the successful event
                 event.reply("Schedule cleared successfully. You are unsubscribed from any reminder alerts.").queue();
             }
         }
